drowsiness-detecction-backend/requirements.txt
absl-py==2.3.1
alembic==1.17.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
arrow==1.4.0
attrs==25.4.0
bcrypt==4.2.1
binaryornot==0.4.4
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cookiecutter==2.6.0
cryptography==46.0.3
cycler==0.12.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.120.0
flatbuffers==25.9.23
flet==0.28.3
flet-cli==0.28.3
flet-desktop==0.28.3
fonttools==4.60.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jax==0.8.0
jaxlib==0.8.0
Jinja2==3.1.6
kiwisolver==1.4.9
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.7
mdurl==0.1.2
mediapipe==0.10.14
ml_dtypes==0.5.3
numpy==2.2.6
oauthlib==3.3.1
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
passlib==1.7.4
pillow==12.0.0
protobuf==4.25.8
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
python-slugify==8.0.4
PyYAML==6.0.3
qrcode==7.4.2
repath==0.9.0
requests==2.32.5
rich==14.2.0
rsa==4.9.1
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
SQLAlchemy==2.0.44
starlette==0.48.0
text-unidecode==1.3
toml==0.10.2
types-python-dateutil==2.9.0.20251008
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchdog==4.0.2
watchfiles==1.1.1
websockets==15.0.1
bcrypt